+++ packages/central-server/app/admin/referenceDataImporter/loaders.js
@@ -717,145 +717,3 @@ export async function procedureTypeLoader(item, { models, pushError }) {
 
   return rows;
 }
-
-// Invoice Price List loader
-function parsePrice(value) {
-  if (value === undefined || value === null || `${value}`.trim() === '') {
-    return null;
-  }
-  const num = Number(value);
-  return Number.isNaN(num) ? undefined : num;
-}
-
-async function initializePriceLists(priceListCodes, models, state, pushError) {
-  const trimmedCodes = priceListCodes.map(c => c.trim());
-  const seen = new Set();
-  const priceListRows = [];
-
-  const existingPriceLists = await models.InvoicePriceList.findAll({
-    where: { code: { [Op.in]: trimmedCodes } },
-  });
-  const existingByCode = new Map(existingPriceLists.map(pl => [pl.code, pl.id]));
-
-  for (const code of priceListCodes) {
-    const trimmedCode = code.trim();
-    if (seen.has(trimmedCode)) {
-      pushError(`duplicate price list code: ${trimmedCode}`);
-      continue;
-    }
-    seen.add(trimmedCode);
-
-    const id = existingByCode.get(trimmedCode) || uuidv4();
-    state.codeToId.set(code, id);
-    priceListRows.push({ model: 'InvoicePriceList', values: { id, code: trimmedCode } });
-  }
-
-  return priceListRows;
-}
-
-function validateAndCollectPrices(item, priceListCodes, invoiceProductId, pushError) {
-  const pricesByCode = new Map();
-
-  for (const code of priceListCodes) {
-    const raw = item[code];
-    const price = parsePrice(raw);
-
-    if (price === undefined) {
-      pushError(
-        `Invalid price value '${raw}' for priceList '${code}' and invoiceProductId '${invoiceProductId}'`,
-      );
-      return null;
-    }
-
-    if (price !== null) {
-      pricesByCode.set(code, price);
-    }
-  }
-
-  return pricesByCode;
-}
-
-async function buildPriceListItems(pricesByCode, invoiceProductId, state, models, pushError) {
-  if (!state.existingItemsMap.has(invoiceProductId)) {
-    const existingItems = await models.InvoicePriceListItem.findAll({
-      where: { invoiceProductId },
-    });
-    const itemsMap = new Map(
-      existingItems.map(item => [`${item.invoicePriceListId}:${item.invoiceProductId}`, item.id]),
-    );
-    state.existingItemsMap.set(invoiceProductId, itemsMap);
-  }
-
-  const productItemsMap = state.existingItemsMap.get(invoiceProductId);
-  const rows = [];
-
-  for (const [code, price] of pricesByCode.entries()) {
-    const invoicePriceListId = state.codeToId.get(code);
-    if (!invoicePriceListId) {
-      pushError(`Could not find InvoicePriceList ID for code '${code}'`);
-      return [];
-    }
-
-    const itemKey = `${invoicePriceListId}:${invoiceProductId}`;
-    const id = productItemsMap.get(itemKey) || uuidv4();
-
-    rows.push({
-      model: 'InvoicePriceListItem',
-      values: { id, invoicePriceListId, invoiceProductId, price },
-    });
-  }
-
-  return rows;
-}
-
-export function invoicePriceListLoader() {
-  const state = {
-    initialized: false,
-    invoiceProductKey: null,
-    priceListCodes: [],
-    codeToId: new Map(),
-    existingItemsMap: new Map(),
-  };
-
-  async function processRow(item, { pushError, models }) {
-    const invoiceProductId = item[state.invoiceProductKey];
-    if (!invoiceProductId) return [];
-
-    const pricesByCode = validateAndCollectPrices(
-      item,
-      state.priceListCodes,
-      invoiceProductId,
-      pushError,
-    );
-    if (!pricesByCode) return [];
-
-    return buildPriceListItems(pricesByCode, invoiceProductId, state, models, pushError);
-  }
-
-  return async (rawItem, { pushError, models }) => {
-    // normalizeItemKeys
-    const item = Object.fromEntries(Object.entries(rawItem).map(([k, v]) => [k?.trim?.() ?? k, v]));
-
-    if (!state.initialized) {
-      const headers = Object.keys(item);
-      const invoiceProductKey = headers.find(h => h.toLowerCase() === 'invoiceproductid');
-
-      if (!invoiceProductKey) {
-        pushError('Missing required column: invoiceProductId');
-        return [];
-      }
-
-      const priceListCodes = headers.filter(h => h !== invoiceProductKey);
-      state.invoiceProductKey = invoiceProductKey;
-      state.priceListCodes = priceListCodes;
-      state.initialized = true;
-
-      const priceListRows = await initializePriceLists(priceListCodes, models, state, pushError);
-      const itemRows = await processRow(item, { pushError, models });
-
-      return [...priceListRows, ...itemRows];
-    }
-
-    return processRow(item, { pushError, models });
-  };
-}
